Fix accept match functionality
- Add more mock functions

+++ src/main/js/match/pending/PendingMatchRepository.js
@@ -1,5 +1,6 @@
 const DbProvider = require('../../DBProvider');
 const featureToggles = require('../../settings/feature-toggles');
+const Optional = require('optional-js');
 const Try = require('try-js');
 
 class PendingMatchRepository {
@@ -10,7 +11,7 @@ class PendingMatchRepository {
 
     getPendingMatches(userId) {
         if (featureToggles.mockDb.enabled) {
-            return Try.of(() => [
+            return Try.success([
                 {
                     "id": "123"
                 }
@@ -20,18 +21,28 @@ class PendingMatchRepository {
         const sql = `SELECT * FROM pending_match
                 WHERE user_id = $(user_id)`;
         const params = { user_id: userId };
-        //return this.db.query(sql, params)
         return Try.of(() => this.db.query(sql, params));
     }
 
     getPendingMatch(pendingMatchId) {
+        if (featureToggles.mockDb.enabled) {
+            return Try.success(Optional.ofNullable({
+                "id": "123"
+            }));
+        }
+
         const sql = `SELECT * FROM pending_match
                 WHERE id = $(pending_match_id)`;
         const params = { pending_match_id: pendingMatchId };
-        return this.db.query(sql, params);
+        return Try.of(() => this.db.query(sql, params))
+            .map(result => Optional.ofNullable(result));
     }
 
     deletePendingMatch(pendingMatchId) {
+        if (featureToggles.mockDb.enabled) {
+            return Try.success();
+        }
+
         const sql = `DELETE FROM pending_match
                 WHERE id = $(pending_match_id)`;
         const params = { pending_match_id: pendingMatchId };
